requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
ruff>=0.4
mypy>=1.10
django-stubs>=5.0
djangorestframework-stubs>=3.14
